requirements.txt
torch>=1.10.0
torchvision>=0.11.0
open_clip_torch>=2.20.0
lpips>=0.1.4
pytorch_msssim>=0.2.1
Pillow>=9.0.0
numpy>=1.21.0
pyyaml>=6.0
tqdm>=4.62.0
